improved deletion condition

+++ src/test/java/CreateOrderTest.java
@@ -97,10 +97,7 @@ public class CreateOrderTest {
     @After
     public void cleanUp(){
 
-        if (accessToken == null) {}
-        else  {
-            UserAPI.deleteUser(accessToken);
-        }
+        if (accessToken != null) {UserAPI.deleteUser(accessToken);}
 
     }
 }
