+++ src/test/java/UpdateUserTest.java
@@ -52,10 +52,7 @@ public class UpdateUserTest {
     @After
     public void cleanUp(){
 
-        if (accessToken == null) {}
-        else  {
-            UserAPI.deleteUser(accessToken);
-        }
+        if (accessToken != null) {UserAPI.deleteUser(accessToken);}
 
     }
 }
