+++ src/test/java/UserLoginTest.java
@@ -65,10 +65,9 @@ public class UserLoginTest {
     @After
     public void cleanUp(){
 
-        if (accessToken == null) {}
-        else  {
-            UserAPI.deleteUser(accessToken);
-        }
+        if (accessToken != null) {UserAPI.deleteUser(accessToken);}
 
     }
-}
+
+    }
+
