+++ src/test/java/UserRegisterTest.java
@@ -57,10 +57,7 @@ public class UserRegisterTest {
     @After
     public void cleanUp(){
 
-        if (accessToken == null) {}
-        else  {
-            UserAPI.deleteUser(accessToken);
-        }
+        if (accessToken != null) {UserAPI.deleteUser(accessToken);}
 
     }
